requirements.txt
anthropic>=0.40.0
mss>=9.0
Pillow>=10.0
pyautogui>=0.9.54
pyperclip>=1.8
pynput>=1.7
pywin32>=306
chromadb>=0.5
sentence-transformers>=3.0
PyPDF2>=3.0
python-docx>=1.0
openpyxl>=3.1
pandas>=2.0
chardet>=5.0
pyyaml>=6.0
python-dotenv>=1.0
imagehash>=4.3
